+++ facade/src/test/scala/react/semanticui/addons/portal/PortalTests.scala
@@ -1,0 +1,18 @@
+package react.semanticui.addons.portal
+
+import utest._
+import japgolly.scalajs.react.test._
+import japgolly.scalajs.react.vdom.html_<^._
+
+object PortalTests extends TestSuite {
+  val tests = Tests {
+    test("render") {
+      val portal = Portal(open = true)(<.div("Abc"))
+      ReactTestUtils.withNewBodyElement { mountNode =>
+        portal.renderIntoDOM(mountNode)
+        println(mountNode.outerHTML)
+        assert(mountNode.outerHTML == """<div></div>""")
+      }
+    }
+  }
+}